% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{add_obs_data}
\alias{add_obs_data}
\title{Given potential outcomes schedule, randomize within block and generate
observed potential outcomes and add them to the schedule.}
\usage{
add_obs_data(dat, p = 0.5, Y0 = "Y0", Y1 = "Y1", blockvar = "B")
}
\arguments{
\item{dat}{Dataframe with a named Y0, Y1, and block column}

\item{p}{Proportion of units treated.  Can be vector and will treat that
proportion in each block, rounded to nearest and with at least 1 treatment
and control unit.}

\item{Y0}{name of Y0 column}

\item{Y1}{name of Y1 column}

\item{blockvar}{name of blocking column.  This column will be converted to a
factor, if it is not already, and the order of p corresponds to the levels
of this factor.}
}
\value{
augmented `dat` with Z and Yobs columns
}
\description{
Given potential outcomes schedule, randomize within block and generate
observed potential outcomes and add them to the schedule.
}
