% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{analysis.combination}
\alias{analysis.combination}
\title{Test for treatment variation across site.
This is a combination test: it uses a liklihood ratio test on model with both systematic and ideosyncratic variation}
\usage{
analysis.combination(df)
}
\description{
Test for treatment variation across site.
This is a combination test: it uses a liklihood ratio test on model with both systematic and ideosyncratic variation
}
