% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIRC_method.R
\name{analysis.FIRC.cov}
\alias{analysis.FIRC.cov}
\title{Covariate adjusted test for cross-site variation}
\usage{
analysis.FIRC.cov(Yobs, Z, B, X, siteID = NULL, data = NULL,
  REML = FALSE, anova = FALSE)
}
\arguments{
\item{X}{Site-level covariate ideally predictive of treatment variation.}

\item{df}{Dataframe to fit.  Needs Z, X, B, and Yobs columns.}
}
\description{
This method fits a FIRC model but also includes a site-level covariate, X,
potentially predictive of treatment impact.
}
\details{
This fits unpooled FIRC models.
}
