% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIRC_method.R
\name{analysis.FIRC}
\alias{analysis.FIRC}
\title{Test for treatment heterogeniety with the FIRC model.}
\usage{
analysis.FIRC(data, REML = FALSE, anova = FALSE)
}
\arguments{
\item{anova}{Use the anova() method to do the test for significance between
the models.  FALSE means do the modified chi-squared test.}
}
\description{
Test for treatment heterogeniety with the FIRC model.
}
