% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoM_method.R
\name{analysis.MoM.unweighted}
\alias{analysis.MoM.unweighted}
\title{This file contains functions implementing the various methods for detecting
and estimating treatment effect variation across sites in a multi-site trial
using methods of moment (MoM) approach.}
\usage{
analysis.MoM.unweighted(df)
}
\description{
We have:
  * Unweighted MoM
  * Weighted MoM
}
