% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstatistic_method.R
\name{analysis.Qstatistic}
\alias{analysis.Qstatistic}
\title{Calculate confidence interval for cross site variation using Q-statistic test inversion.}
\usage{
analysis.Qstatistic(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  alpha = 0.05,
  calc.CI = FALSE
)
}
\arguments{
\item{Yobs}{(outcome)}

\item{Z}{(binary treatment 0/1)}

\item{B}{(site id)}

\item{data}{Dataframe with outcome, treatment, and blocking factor.}

\item{alpha}{The level of the test.  The CI will be a 1-2alpha confidence interval.}
}
\description{
Code taken from prior work of Catherine Armstrong.
}
