% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{analysis.systematic}
\alias{analysis.systematic}
\title{Test for cross site variation by testing for a covariate predictive of variation.}
\usage{
analysis.systematic(df)
}
\description{
Comment: Should this model allow for random tx variation or no?  (Currently it does not.)
}
