% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{analysis.systematic.RTx}
\alias{analysis.systematic.RTx}
\title{Systematic test with the random ideosyncratic component}
\usage{
analysis.systematic.RTx(df)
}
\description{
Systematic test with the random ideosyncratic component
}
