% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIRC_method.R
\name{analysis_idio}
\alias{analysis_idio}
\alias{analysis_systematic.RTx}
\alias{analysis_combination}
\title{Test for cross site variation using a liklihood ratio test vs. a model with no random slope (but random intercept).
This is the RIRC version of such a test.}
\usage{
analysis_idio(df)

analysis_systematic.RTx(df)

analysis_combination(df)
}
\description{
This is a combination test: it uses
a liklihood ratio test on model with both systematic and ideosyncratic
variation
}
