% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{analysis_Qstatistic_stat}
\alias{analysis_Qstatistic_stat}
\alias{analysis_Qstatistic}
\title{Calculate confidence interval for cross site variation using Q-statistic test
inversion.}
\usage{
analysis_Qstatistic_stat(
  ATE_hat,
  SE_hat,
  alpha = 0.05,
  calc_CI = FALSE,
  calc_optim_pvalue = FALSE,
  verbose = FALSE,
  tolerance = 1e-05,
  mean_method = c("weighted", "raw")
)

analysis_Qstatistic(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  alpha = 0.05,
  calc_CI = FALSE
)
}
\arguments{
\item{ATE_hat}{List of estimated ATEs, one for each site.}

\item{SE_hat}{List of associated estimated SEs, assumed to be known.}

\item{alpha}{The level of the test.  The CI will be a 1-2alpha confidence
interval.}

\item{calc_CI}{Logical, TRUE means calculate the confidence interval using
test-inversion.}

\item{calc_optim_pvalue}{TRUE if estimate cross site variation by maximizing
p-value (a type of Hodges-Lehman estimate).  FALSE means uses
Konstantopoulos and Hedges direct formula. (TRUE requires calc_CI also be
set to TRUE.)  Default of FALSE.  FALSE recommended.}

\item{verbose}{Print out a few diagnostics as the CI is being calculated, for
those curious.}

\item{tolerance}{For optimization and search for CI endpoints, how close to
optimal do we want?}

\item{mean_method}{How to calculate the mean effect.  'weighted' is a
precision weighted mean. 'raw' is the simple mean.  Or pass a number which
will be used as the mean.}

\item{Yobs}{(outcome)}

\item{Z}{(binary treatment 0/1)}

\item{B}{(site id)}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified
by siteID).}
}
\value{
List of several estimated quantities: cross-site variation (standard
  deviation, not variance), p-value for presence of any variation, flag of
  whether the p-value is less than passed alpha, the Q statistic itself, and
  the confidence interval.
}
\description{
Two versions of this, one that takes raw data (in which case a linear model
is fit to it to get individual site impact estimates), and a second, '_stat',
that takes precalculated point estimates and standard errors.
}
\details{
The first simply calculates ATE_hat and SE_hat, and then calls the second.

Cross site variation, by default, is estimated by method of moments technique
of Konstantopoulos and Hedges). One can set an alternate method for
estimating cross-site variation via using optim() to find the point estimate
corresponding to maximizing the p-value (this differs slightly from direct
method of moment); this latter approach requires calculating the confidence
interval using test inversion as well.

Code based on prior work of Catherine Armstrong, which was then augmented to
do binary search for endpoints for the test inversion procedure to generate
confidence intervals.
}
\examples{
analysis_Qstatistic_stat( rnorm(30, sd=0.1), rep(0.1,30), calc_CI = TRUE )
}
