% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_systematic.R
\name{analysis_systematic}
\alias{analysis_systematic}
\title{Test for cross site variation by testing for a covariate predictive of variation.}
\usage{
analysis_systematic(df)
}
\arguments{
\item{df}{Dataframe to analyze.}
}
\description{
Comment: Should this model allow for random tx variation or no?  (Currently it does not.)
}
