% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisiteresult.R
\name{as.data.frame.multisiteresult}
\alias{as.data.frame.multisiteresult}
\alias{multisiteresult}
\title{Cast qstat result to data.frame}
\usage{
\method{as.data.frame}{multisiteresult}(x)
}
\arguments{
\item{x}{the multisiteresult object to covert}
}
\description{
Cast qstat result to data.frame
}
\seealso{
Other multisiteresult: 
\code{\link{is.multisiteresult}()},
\code{\link{print.multisiteresult}()}
}
\concept{multisiteresult}
