% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{as.data.frame.qstatresult}
\alias{as.data.frame.qstatresult}
\alias{qstatresult}
\title{Cast qstat result to data.frame}
\usage{
\method{as.data.frame}{qstatresult}(x)
}
\arguments{
\item{x}{the qstatresult object to covert}
}
\description{
Cast qstat result to data.frame
}
\seealso{
Other qstatresult: 
\code{\link{is.qstatresult}()},
\code{\link{print.qstatresult}()}
}
\concept{qstatresult}
