% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blkvar.R
\docType{package}
\name{blkvar}
\alias{blkvar}
\title{\code{blkvar} package}
\description{
Blocked Design Variance Estimation
}
\details{
See the README on
\href{https://github.com/lmiratrix/blkvar#readme}{GitHub}
}
