% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_var_est_func.R
\name{block.data}
\alias{block.data}
\title{Block table function}
\usage{
block.data(Y, Z, B, data = NULL)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{alternatively is matrix of Y,Z,B}
}
\description{
Function that returns a summary of block level estimates.
See also `calc.summary.stats` (which does the same thing with slightly different interface).
}
