% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{block_data_sim}
\alias{block_data_sim}
\title{Table of data for simulations}
\usage{
block_data_sim(Y, Z, B, p.mat, data = NULL)
}
\arguments{
\item{Y}{vector of all outcomes}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{B}{block ids}

\item{p.mat}{matrix with first column B,second column prop treated in that block, p}

\item{data}{alternatively is matrix of Y,Z,B}
}
\description{
Function that returns a summary of block level true values for sims.
}
