% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{block_distn}
\alias{block_distn}
\title{Generate piecewise uniform distribution with a mean of n.bar}
\usage{
block_distn(J, n.bar, size.ratio)
}
\description{
Generate piecewise uniform distribution with a mean of n.bar
}
