% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{block_estimator}
\alias{block_estimator}
\title{Block variance estimation function.}
\usage{
block_estimator(
  Yobs,
  Z,
  B,
  data = NULL,
  method = c("hybrid_m", "hybrid_p", "plug_in_big", "rct_yes_all", "rct_yes_small",
    "rct_yes_mod_all", "rct_yes_mod_small"),
  throw.warnings = TRUE
)
}
\arguments{
\item{Yobs}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors}

\item{method}{is method of variance estimation, defauly "hybrid_m"}

\item{throw.warnings}{TRUE means throw warnings if the hybrid estimators are}
}
\description{
This function takes observed data and returns a treatment effect estimate,
variance estimate and some summary info about the block structure.
}
