% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCT_yes_formula.R
\name{calc.RCT.Yes.SE}
\alias{calc.RCT.Yes.SE}
\title{Implementation of the formula found in RCT-YES documentation.}
\usage{
calc.RCT.Yes.SE(sum_tab, method = c("finite", "superpop",
  "superpop.original"), weight = c("individual", "site"))
}
\arguments{
\item{sum_tab}{Table of summary statistics by block, from, e.g.,
`block.data()`}

\item{method}{finite, superpop, or superpop2 to give SEs that either capture
uncertainty due to targeting a superpopulation quantity or not.}

\item{weight}{Individual weight (i.e., number of individuals in each block)
or site weight (average block estimates).}
}
\description{
This can handle a superpopulation model, non-clustered but blocked.
}
\details{
Taken from page 83 of Shochet RCT-YES paper (eq 6.25).

The `superpop2` variant is a modification of the original, pulling the
weights from inside the squared term to outside.  If the weights are not all
1, this can make a difference.
}
