% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_method.R
\name{calc.summary.stats.formula}
\alias{calc.summary.stats.formula}
\title{Summarise data by block.}
\usage{
calc.summary.stats.formula(formula, data = NULL, siteID = NULL,
  add.neyman = FALSE)
}
\arguments{
\item{formula}{Formula of form Yobs ~ Z * B (ONLY)}

\item{siteID}{If not null, name of siteID that has randomization blocks
nested inside.}

\item{dat}{Dataframe with defined Yobs, Z, and B variables.}
}
\value{
dataframe with summary statistics by block
}
\description{
Given dataframe or list of variables, return table of stats for each
randomization block.
}
