% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{calc.summary.stats.oracle}
\alias{calc.summary.stats.oracle}
\title{Summarise simulation data by block (where we know both Y0 and Y1)}
\usage{
calc.summary.stats.oracle(dt)
}
\arguments{
\item{dat}{Dataframe with defined Y0, Y1, and blk variables.}
}
\value{
dataframe with summary statistics by block
}
\description{
Summarise simulation data by block (where we know both Y0 and Y1)
}
