% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_var_est_func.R
\name{calc.summary.stats}
\alias{calc.summary.stats}
\title{Summarise data by block.}
\usage{
calc.summary.stats(dat)
}
\arguments{
\item{dat}{Dataframe with defined Yobs, Z, and blk variables.}
}
\value{
dataframe with summary statistics by block
}
\description{
Summarise data by block.
}
