% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{calc_cross_site_variation}
\alias{calc_cross_site_variation}
\title{Calculate cross site variation}
\usage{
calc_cross_site_variation(Y, SE)
}
\arguments{
\item{Y}{List of estimated impacts for sites}

\item{SE}{List of standard errors for the estimated impacts.}
}
\value{
Grand mean, Q statistic, and estimated tau (cross site variation).
}
\description{
Given a set of impact estimates and associated standard errors, use method of
moments to calculate the implied cross site variation, assuming the standard
errors are correct (and that site-specific impacts are independent of
precision).
}
\details{
For discussion of this approach, see pg 241 of Handbook of Quantitative
Methods for Educational Research, Timothy Teo (Ed.)
}
