% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_site_power_calc.R
\name{calc_MDESSD}
\alias{calc_MDESSD}
\title{Calculate MDESSD for cross-site variation}
\usage{
calc_MDESSD(J, n.bar, p.tx = 0.5, ICC = 0.5, K = 0, R2.X = 0, alpha = 0.05)
}
\arguments{
\item{J}{Number of sites}

\item{n.bar}{average people per site}

\item{p.tx}{proportion treated, Default: 0.5}

\item{ICC}{Control-side ICC, Default: 0.5}

\item{K}{Number of individual covariates used to explain control-side variation, Default: 0}

\item{R2.X}{Explanatory power of these individual level covariates, Default: 0}

\item{alpha}{Significance level, Default: 0.05}
}
\value{
MDESSD for given scenario.
}
\description{
Calculate MDESSD (minimum detectable effect size for cross site standard deviation of effects
assuming a ratio test.  This uses work and results found in Appendix A of Bloom and Spybrook
}
\seealso{
calc_power
}
