% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_power.R
\name{calc_power}
\alias{calc_power}
\title{Calculate power for cross-site variation}
\usage{
calc_power(J, n.bar, tau, p.tx = 0.5, ICC = 0.5, K = 0, R2.X = 0, alpha = 0.05)
}
\arguments{
\item{J}{Number of sites}

\item{n.bar}{average people per site}

\item{tau}{Cross site VARIANCE of site-level ATEs.}

\item{p.tx}{proportion treated, Default: 0.5}

\item{ICC}{Control-side ICC, Default: 0.5}

\item{K}{Number of individual covariates used to explain control-side variation, Default: 0}

\item{R2.X}{Explanatory power of these individual level covariates, Default: 0}

\item{alpha}{Significance level, Default: 0.05}
}
\value{
Power
}
\description{
Calculate power for detecting cross site variation with a ratio
  test.  This uses work and results in Appendix A of Bloom and Spybrook.
}
\examples{
# Example in the text
calc_power( 80, 60, 0.02, p.tx=0.6, ICC=0.2, K = 0, R2.X=0.25 )

}
