% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_stats.R
\name{calc_summary_stats}
\alias{calc_summary_stats}
\title{Summarise data by block.}
\usage{
calc_summary_stats(Yobs, Z, B, data = NULL, siteID = NULL, add.neyman = FALSE)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{B}{block ids}

\item{data}{Dataframe with defined Yobs, Z, and B variables.}

\item{siteID}{If not null, name of siteID that has randomization blocks nested inside.}

\item{add.neyman}{TO ADD}
}
\value{
dataframe with summary statistics by block
}
\description{
Given dataframe or list of variables, return table of stats for each
randomization block.
}
