% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare.methods}
\alias{compare.methods}
\title{Block variance method comparison function}
\usage{
compare.methods(Y, Z, B, data = NULL)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors.  The columns
of the data matrix have to be in the order of Y, Z, B as column 1, 2, 3.}
}
\description{
This function calculates the point estimates and SE estimates for a variety
of the blocked designs.
}
