% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{compare.methods.sims}
\alias{compare.methods.sims}
\title{Block variance method comparison function specifically for sim values.}
\usage{
compare.methods.sims(Y, Z, B, data = NULL, include.MLM = TRUE,
  include.RCTYes = TRUE, include.LM = TRUE)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors.  The columns
of the data matrix have to be in the order of Y, Z, B as column 1, 2, 3.}

\item{include.MLM}{Include MLM estimators}

\item{include.RCTYes}{Include RCTYes estimators}

\item{include.LM}{Include Linear Model-based estimators (including Huber-White SEs, etc.)}
}
\description{
This function calculates the point estimates and SE estimates for a variety
of the blocked designs.
}
