% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{compare.methods.true.val}
\alias{compare.methods.true.val}
\title{Function to compare estimators using the true value (whole science table)}
\usage{
compare.methods.true.val(Y, Z, B, p.mat, data = NULL)
}
\arguments{
\item{Y}{vector of all outcomes}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{B}{block ids}

\item{p.mat}{matrix with first column Blk_ID,second column prop treated in that block, p}

\item{data}{alternatively is matrix of Y,Z,B}
}
\description{
Function that returns some variance function estimates based on all potential outcomes.
}
