% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{Block variance method comparison function}
\usage{
compare_methods(Y, Z, B, siteID = NULL, data = NULL,
  include.block = TRUE, include.MLM = TRUE, include.RCTYes = TRUE,
  include.LM = TRUE, include.RCTYesBlended = FALSE,
  include.method.characteristics = FALSE)
}
\arguments{
\item{Y}{vector observed outcomes (or column name in data)}

\item{Z}{vector of assignment indicators (1==treated) (or column name in data)}

\item{B}{vector of block ids (or column name in data)}

\item{siteID}{site ids (variable name as string if data frame passed) (if randomization blocks are nested in site).}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified by siteID).}

\item{include.block}{Include the Pashley blocking variants.}

\item{include.MLM}{Include MLM estimators}

\item{include.RCTYes}{Include RCTYes estimators}

\item{include.LM}{Include Linear Model-based estimators (including
Huber-White SEs, etc.)}

\item{include.RCTYesBlended}{Include RCTYes estimator applied to small block
and classic Neyman to large blocks.}

\item{include.method.characteristics}{Include details of the methods (target estimands and sampling framework assumed).}
}
\description{
This function calculates the point estimates and SE estimates for a variety
of the blocked designs.
}
