% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_oracle.R
\name{compare_methods_oracle}
\alias{compare_methods_oracle}
\title{Function to compare estimators using the true values (i.e., whole science
table).  This is for simulaton studies where we know all the potential
outcomes.}
\usage{
compare_methods_oracle(Y0, Y1, B, data = NULL, p_mat = NULL)
}
\arguments{
\item{Y0}{Vector of control potential outcomes.}

\item{Y1}{Vector of treatment potential outcomes}

\item{B}{block ids}

\item{data}{alternatively is matrix of Y,Z,B}

\item{p_mat}{matrix with two columns, one row per block.  First column is B,
second column prop treated in that block, ("p")}
}
\description{
Function that returns some variance function estimates based on all potential
outcomes.
}
