% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods_variation}
\alias{compare_methods_variation}
\title{Compare different estimates of cross site variation.}
\usage{
compare_methods_variation(Yobs, Z, B, data = NULL,
  include.testing = TRUE, long.results = FALSE)
}
\arguments{
\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors.  The columns
of the data matrix have to be in the order of Y, Z, B as column 1, 2, 3.}

\item{long.results}{TRUE means each estimator gets a line in a data.frame.  FALSE gives all as columns in a 1-row dataframe.}
}
\description{
Given a dataframe, use the different methods to pull out point estimates and
(if desired) pvalues and return them all.
}
