% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_MLM_methods.R
\name{compare_MLM_methods}
\alias{compare_MLM_methods}
\title{Compare different MLM methods}
\usage{
compare_MLM_methods(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  control.formula = NULL
)
}
\arguments{
\item{Yobs}{vector observed outcomes (or column name in data)}

\item{Z}{vector of assignment indicators (1==treated) (or column name in data)}

\item{B}{vector of block ids (or column name in data)}

\item{siteID}{site ids (variable name as string if data frame passed) (if randomization blocks are nested in site).}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified by siteID).}

\item{control.formula}{What variables to control for, in the form of "~ X1 + X2".}
}
\description{
Compare different MLM methods
}
