% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_oracle.R
\name{describe_data}
\alias{describe_data}
\title{Describe an oracle (simulated) dataset.}
\usage{
describe_data(data, Y0 = "Y0", Y1 = "Y1", Z = "Z", sid = "sid")
}
\arguments{
\item{data}{Dataframe with defined Y0, Y1 (the potential outcomes), Z (sample
treatment assignment vector), and sid (a categorical blocking variable).}

\item{Y0}{name of Y0 column}

\item{Y1}{name of Y1 column}

\item{Z}{example vector of treatment assignment that allows for calculation
of proportion treated and so forth.}

\item{sid}{Name of the block ID column in the dataframe.}
}
\value{
(invisible) the summary stats by block.  Side effect of printing to
  screen the description of the data.
}
\description{
This utility function gives various summary statistics (at the block level)
for a dataset with all potential outcomes given.
}
