% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plot.R
\name{diag_plot}
\alias{diag_plot}
\title{Make diagnostic plot}
\usage{
diag_plot(Y, Z, B, data = NULL)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors}
}
\description{
Function that plots variance estimates versus size of treatment group.
}
