% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_method.R
\name{estimate.ATE.design.based}
\alias{estimate.ATE.design.based}
\title{Implementation of the formula found in RCT-YES documentation.}
\usage{
estimate.ATE.design.based(sum_tab, siteID = NULL, method = c("finite",
  "superpop", "superpop.original"), weight = c("individual", "site"))
}
\arguments{
\item{sum_tab}{Table of summary statistics by block, from, e.g.,
`block.data()`}

\item{siteID}{Vector of site IDs if there are randomization blocks nested in
site that should be aggregated (will change results for site weighting only).}

\item{method}{finite, superpop, or superpop2 to give SEs that either capture
uncertainty due to targeting a superpopulation quantity or not.}

\item{weight}{Individual weight (i.e., number of individuals in each block)
or site weight (average site estimates (which will be considered block
estimates if siteID is null)).}
}
\value{
dataframe with calculated impacts and standard errors.
}
\description{
This can handle a superpopulation model, non-clustered but blocked.
}
\details{
Taken from page 83 of Shochet RCT-YES paper (eq 6.25).

The `superpop` variant is a modification of the original 'superpop.original',
pulling the weights from inside the squared term to outside. This method was
suggested in personal correspondance with Schochet.  If the weights are not
all 1, this can make a difference.
}
