% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIRC_method.R
\name{estimate.ATE.FIRC}
\alias{estimate.ATE.FIRC}
\alias{estimate.ATE.FIRC.pool}
\title{Fit the FIRC model to estimate ATE across sites as well as cross site treatment variation.}
\usage{
estimate.ATE.FIRC(Yobs, Z, sid, data = NULL, REML = FALSE)

estimate.ATE.FIRC.pool(Yobs, Z, sid, data = NULL)
}
\value{
List of things including a pvalue for cross-site variation
}
\description{
Acknowledgement: taken and adapted from Catherine's weiss.tau() method.

For testing for cross site variation, this will adjust the ratio stat p-value to account for the mixture of
chi-squared distributions.
}
