% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIRC_method.R
\name{estimate.ATE.RIRC}
\alias{estimate.ATE.RIRC}
\alias{estimate.ATE.RICC}
\title{Estimate the ATE using Random-Intercept, Random-Coefficient (RIRC) Models}
\usage{
estimate.ATE.RIRC(Yobs, Z, sid, data = NULL, REML = FALSE,
  include.testing = TRUE)

estimate.ATE.RICC(Yobs, Z, sid, data = NULL, REML = FALSE)
}
\description{
This model has a single treatment coefficient, and a random intercept for the
site control average. So it is analogous to a fixed effect model, but with a
random effect.
}
\details{
There is no test for cross site variation for this method, since we assume none.
}
