% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_method.R
\name{estimate_ATE_design_based_adjusted}
\alias{estimate_ATE_design_based_adjusted}
\title{Adjusted design based estimator for ATE}
\usage{
estimate_ATE_design_based_adjusted(
  formula,
  control.formula,
  data,
  siteID = NULL,
  method = c("finite", "superpop", "superpop.adj"),
  weight = c("individual", "site")
)
}
\arguments{
\item{formula}{The formula argument must be of the form outcome ~ treatment:block_id.}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}

\item{data}{Dataframe with all needed variables.}

\item{method}{What form of standard errors (finite, superpop, or adjusted superpop)}

\item{weight}{Site or individual weighting.}
}
\value{
Estimates of ATE along with SEs.
}
\description{
Given two covariates, X1 and X2, calculate adjusted ATE estimates for the
multisite trial.  This uses the formula from the Schochet RCT Yes technical
document.
}
