% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIRC_method.R
\name{estimate_ATE_RIRC_pool}
\alias{estimate_ATE_RIRC_pool}
\title{Fit Random-intercept, random-slope model}
\usage{
estimate_ATE_RIRC_pool(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  include.testing = TRUE,
  control.formula = NULL
)
}
\description{
This is a wrapper for a simpler lmer call.
}
