% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_var_est_func.R
\name{fitdata.sumtable}
\alias{fitdata.sumtable}
\title{Block variance estimation function.}
\usage{
fitdata.sumtable(data.table, method = c("hybrid_m", "hybrid_p",
  "plug_in_big"), throw.warnings = TRUE)
}
\arguments{
\item{data.table}{Summary statistics of blocks in a dataset.  In particular, the output of block.data method.}

\item{method}{is method of variance estimation, defauly "hybrid_m"}

\item{throw.warnings}{TRUE means throw warnings if the hybrid estimators are breaking down due to violation of assumptions.}
}
\description{
This function takes the block-level summary statistics of a dataset and returns a treatment effect estimate,
variance estimate and some summary info about the block structure.
}
