% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{fixed_effect_estimators}
\alias{fixed_effect_estimators}
\title{Fixed effect linear regression}
\usage{
fixed_effect_estimators(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  block.stats = NULL,
  control_formula = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{Dataframe of the data to analyse.}

\item{block.stats}{Table of precomputed block-level statistics (optional, for speed concerns; this gets precomputed in compare_methods).}

\item{control_formula}{The control_formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}
}
\value{
Dataframe of results for different estimators.
}
\description{
Fit regressions with block fixed effects (and a single treatment parameter).
Calculate standard errors in a variety of ways.
}
\details{
For the club sandwich estimation ("Club"), see code proposed at
https://www.jepusto.com/handmade-clubsandwich/
}
