% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_blocked_data.R
\name{form_blocks_from_continuous}
\alias{form_blocks_from_continuous}
\title{Cut continuous covariate into similar-valued blocks}
\usage{
form_blocks_from_continuous(
  X,
  method = c("small", "pair", "big", "none"),
  num.blocks
)
}
\arguments{
\item{X}{covariate vector to block on}

\item{method}{How to block.  "small"  "pair" makes matched pairs. "big" makes as many
blocks of at least 4 as possible, trying to keep blocks small. "none" makes
a single block.}

\item{num.blocks}{If method is small, how many blocks to attempt to make}
}
\value{
Vector with one element per element of `X`
}
\description{
Make blocks out of a continuous covariate by cutting it into pieces that are
ideally relatively homogenous. This will create a bunch of blocks based on
passed covariate and return a factor vector of those blocks.
}
\examples{
table( form_blocks_from_continuous( 1:22, method="small" ) )
table( form_blocks_from_continuous( 1:22, method="big", num.blocks=3 ) )
table( form_blocks_from_continuous( 1:22, method="pair" ) )
table( form_blocks_from_continuous( 1:22, method="none" ) )
table( form_blocks_from_continuous( 1:22, method="big" ) )
}
