% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{gen.dat.model.no.cov}
\alias{gen.dat.model.no.cov}
\title{Generate multilevel data with no covariates.  A simplification of function
above.}
\usage{
gen.dat.model.no.cov(n.bar = 16, J = 30, p = 0.5, gamma.00, gamma.10,
  tau.00, tau.01, tau.11, sigma2.e, variable.n = TRUE,
  variable.p = FALSE, return.sites = FALSE, finite.model = FALSE,
  size.impact.correlate = 0, proptx.impact.correlate = 0,
  correlate.strength = 0.75, size.ratio = 1/3, verbose = FALSE)
}
\arguments{
\item{n.bar}{average site size, Default: 10}

\item{J}{number sites, Default: 30}

\item{p}{prop treated, Default: 0.5}

\item{gamma.00}{The mean control outcome, Default: 0}

\item{gamma.10}{The ATE, Default: 0.2}

\item{size.impact.correlate}{Takes values of -1, 0, or 1: Are site impacts
negatively correlated, uncorrelated, or positively correlated with site
size?}

\item{proptx.impact.correlate}{Takes values of -1, 0, or 1: Are proportion
of units treated negatively correlated, uncorrelated, or positively
correlated with site size?}

\item{correlate.strength}{In [0,1], and describes how correlated the ranking
of site impacts will be with proptx and site size, if they are set to be
correlated.}

\item{verbose}{Say stuff while maing data?, Default: FALSE}
}
\value{
multisite data with at least 2 tx and 2 co units in each block.
}
\description{
Generate multilevel data with no covariates.  A simplification of function
above.
}
