% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{gen.dat.model}
\alias{gen.dat.model}
\title{Generate multilevel data from a model}
\usage{
gen.dat.model(n.bar = 10, J = 30, p = 0.5, gamma.00, gamma.01, gamma.10,
  gamma.11, tau.00, tau.01, tau.11, sigma2.e, variable.n = TRUE,
  return.sites = FALSE, verbose = FALSE)
}
\arguments{
\item{n.bar}{PARAM_DESCRIPTION, Default: 10}

\item{J}{PARAM_DESCRIPTION, Default: 30}

\item{p}{PARAM_DESCRIPTION, Default: 0.5}

\item{gamma.00}{PARAM_DESCRIPTION}

\item{gamma.01}{PARAM_DESCRIPTION}

\item{gamma.10}{PARAM_DESCRIPTION}

\item{gamma.11}{PARAM_DESCRIPTION}

\item{tau.00}{PARAM_DESCRIPTION}

\item{tau.01}{PARAM_DESCRIPTION}

\item{tau.11}{PARAM_DESCRIPTION}

\item{sigma2.e}{PARAM_DESCRIPTION}

\item{variable.n}{PARAM_DESCRIPTION, Default: TRUE}

\item{return.sites}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Given a 2-level model, generate data to specifications
}
