% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{gen.dat.no.cov}
\alias{gen.dat.no.cov}
\title{Simplified version of gen.dat() with no X covariate.}
\usage{
gen.dat.no.cov(n.bar = 10, J = 30, p = 0.5, tau.11.star = 0.3,
  ICC = 0.7, gamma.00 = 0, gamma.10 = 0.2, verbose = FALSE,
  variable.n = TRUE, control.sd.Y1 = TRUE, ...)
}
\arguments{
\item{n.bar}{PARAM_DESCRIPTION, Default: 10}

\item{J}{PARAM_DESCRIPTION, Default: 30}

\item{p}{PARAM_DESCRIPTION, Default: 0.5}

\item{tau.11.star}{PARAM_DESCRIPTION, Default: 0.3}

\item{ICC}{desired ICC, Default: 0.7}

\item{gamma.00}{PARAM_DESCRIPTION, Default: 0}

\item{gamma.10}{PARAM_DESCRIPTION, Default: 0.2}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{variable.n}{PARAM_DESCRIPTION, Default: TRUE}

\item{control.sd.Y1}{Make correlation of random intercept and random slope
such that the variance of the Y1s is 1.0, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
Dataframe of individual data from a MLM DGP.
}
\description{
Generate fake data for simulation studies
}
