% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{gen.dat}
\alias{gen.dat}
\title{FUNCTION_TITLE}
\usage{
gen.dat(n.bar = 10, J = 30, p = 0.5, tau.11.star = 0.3, rho2.0X = 0.1,
  rho2.1X = 0.5, ICC = 0.7, gamma.00 = 0, gamma.10 = 0.2,
  verbose = FALSE, zero.corr = FALSE, ...)
}
\arguments{
\item{n.bar}{average site size, Default: 10}

\item{J}{number sites, Default: 30}

\item{p}{prop treated, Default: 0.5}

\item{tau.11.star}{PARAM_DESCRIPTION, Default: 0.3}

\item{rho2.0X}{PARAM_DESCRIPTION, Default: 0.1}

\item{rho2.1X}{PARAM_DESCRIPTION, Default: 0.5}

\item{ICC}{PARAM_DESCRIPTION, Default: 0.7}

\item{gamma.00}{PARAM_DESCRIPTION, Default: 0}

\item{gamma.10}{PARAM_DESCRIPTION, Default: 0.2}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{zero.corr}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{Further parameters passed to gen.dat.model()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
