% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{generate.individuals.from.blocks}
\alias{generate.individuals.from.blocks}
\title{Function to generate individual-level data from a list of block sizes and
block characteristics.}
\usage{
generate.individuals.from.blocks(n_k, alpha = 0, beta = 0, sigma_c = 1,
  sigma_t = 1, corr = 1, exact = FALSE)
}
\arguments{
\item{n_k}{Vector of block sizes}

\item{alpha}{Vector of (expected) means of the control potential outcome}

\item{beta}{Vector of the block-level Average Treatment Effects}

\item{sigma_c}{Standard deviation of the control potential outcomes.}

\item{sigma_t}{Standard deviation of the treatment potential outcomes}

\item{corr}{Correlation of the potential outcomes.}

\item{exact}{If TRUE generate data so we match the desired means and moments
exactly.  False means pull from bivariate normal.}
}
\value{
Matrix of the potential outcomes and block ids.
}
\description{
This generates potential outcomes by sampling from the specified bivariate
normal distributions within each block.
}
