% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dat.R
\name{generate_multilevel_data_model}
\alias{generate_multilevel_data_model}
\alias{generate_multilevel_data}
\alias{generate_multilevel_data_no_cov}
\title{Generate multilevel data from a model}
\usage{
generate_multilevel_data_model(
  n.bar = 10,
  J = 30,
  p = 0.5,
  gamma.00,
  gamma.01,
  gamma.10,
  gamma.11,
  tau.00,
  tau.01,
  tau.11,
  sigma2.e,
  sigma2.W = 1,
  beta.X = NULL,
  sigma2.mean.X = 0,
  variable.n = TRUE,
  variable.p = FALSE,
  cluster.rand = FALSE,
  return.sites = FALSE,
  finite.model = FALSE,
  size.impact.correlate = 0,
  proptx.impact.correlate = 0,
  correlate.strength = 0.75,
  size.ratio = 1/3,
  verbose = FALSE
)

generate_multilevel_data(
  n.bar = 10,
  J = 30,
  p = 0.5,
  tau.11.star = 0.3,
  rho2.0W = 0.1,
  rho2.1W = 0.5,
  ICC = 0.7,
  gamma.00 = 0,
  gamma.10 = 0.2,
  verbose = FALSE,
  zero.corr = FALSE,
  ...
)

generate_multilevel_data_no_cov(
  n.bar = 10,
  J = 30,
  p = 0.5,
  tau.11.star = 0.3,
  ICC = 0.7,
  gamma.00 = 0,
  gamma.10 = 0.2,
  verbose = FALSE,
  variable.n = TRUE,
  control.sd.Y1 = TRUE,
  ...
)
}
\arguments{
\item{n.bar}{average site size, Default: 10}

\item{J}{number sites, Default: 30}

\item{p}{prop treated, Default: 0.5}

\item{gamma.00}{The mean control outcome, Default: 0}

\item{gamma.01}{Coefficient for W to site control mean}

\item{gamma.10}{The ATE, Default: 0.2}

\item{gamma.11}{Coefficient for W to treatment impact}

\item{tau.00}{Variance of site conrol means}

\item{tau.01}{Correlation of treatment impact and mean site outcome under control}

\item{tau.11}{Treatment impact variance}

\item{sigma2.e}{Residual standard error}

\item{sigma2.W}{The variation of the site-level covariate.}

\item{beta.X}{Coefficient for the individual-level X covariate.  NA means no covariate.}

\item{sigma2.mean.X}{How much the individual-level X covariate means vary across site.}

\item{variable.n}{Allow n to vary around n.bar, Default: TRUE}

\item{variable.p}{Should the proportion of units treated in each site vary?  Yes/No.}

\item{cluster.rand}{TRUE means cluster-randomized.  FALSE means randomized within site.}

\item{return.sites}{Return sites, not individual students, Default: FALSE}

\item{finite.model}{If TRUE use a canonical set of random site effects.  When TRUE this method will save the multivariate normal draw and reuse it in subsequent calls to generate_multilevel_data_model until a call with a different J is made.  Recommended to use FALSE.}

\item{size.impact.correlate}{Takes values of -1, 0, or 1: Are site impacts
negatively correlated, uncorrelated, or positively correlated with site
size?}

\item{proptx.impact.correlate}{Takes values of -1, 0, or 1: Are proportion
of units treated negatively correlated, uncorrelated, or positively
correlated with site size?}

\item{correlate.strength}{In [0,1], and describes how correlated the ranking
of site impacts will be with proptx and site size, if they are set to be
correlated.}

\item{size.ratio}{The degree to which the site sizes should vary, if they should vary.}

\item{verbose}{Say stuff while maing data?, Default: FALSE}

\item{tau.11.star}{Total amount of cross site treatment variation}

\item{rho2.0W}{Explanatory power of W for control outcomes, Default: 0.1}

\item{rho2.1W}{Explanatory power of W for average treatment impact, Default:
0.5}

\item{ICC}{The ICC, Default: 0.7}

\item{zero.corr}{TRUE means treatment impact and mean site outcome are not
correlated.  TRUE means they are negatively correlated to make the variance
of the treatment group 1, Default: FALSE}

\item{...}{Further parameters passed to generate_multilevel_data_model()}

\item{control.sd.Y1}{Make correlation of random intercept and random slope}
}
\value{
Dataframe of individual level data (unless return.sites=TRUE)!  Dataframe has treatment column, outcome column, covariates, and block IDs.
}
\description{
Given a 2-level model, generate data to specifications
}
\section{Functions}{
\itemize{
\item \code{generate_multilevel_data}: Wrapper for generate_multilevel_data_model that rescales parameters to make standardization easier.

\item \code{generate_multilevel_data_no_cov}: Simplified version of generate_multilevel_data() with no W covariate.
}}

