% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ATE_FIRC.R
\name{get_EB_estimates}
\alias{get_EB_estimates}
\title{Get empirical Bayes estimates from fit FIRC model}
\usage{
get_EB_estimates(x)
}
\arguments{
\item{x}{A FIRC model multisiteresult object.}
}
\value{
Tibble with site IDs and estimated site average impacts.
}
\description{
Get empirical Bayes estimates from fit FIRC model
}
