% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{hybrid_m_small}
\alias{hybrid_m_small}
\title{Aggregated matched pairs type variance estimator}
\usage{
hybrid_m_small(data.small)
}
\arguments{
\item{data.small}{data frame containg info for small blocks}
}
\description{
Function to calculate hybrid_m variance estimator.
}
