% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_var_est_func.R
\name{hybrid_p}
\alias{hybrid_p}
\title{Pooled matched pairs type variance estimator}
\usage{
hybrid_p(data.small)
}
\arguments{
\item{data.small}{data frame containg info for small blocks}
}
\description{
Function to calculate hybrid_p variance estimator.
}
