% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{hybrid_p_small}
\alias{hybrid_p_small}
\title{Pooled matched pairs type variance estimator}
\usage{
hybrid_p_small(data.small)
}
\arguments{
\item{data.small}{data frame containg info for small blocks}
}
\description{
Function to calculate hybrid_p variance estimator.
}
