% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacted_linear_estimators.R
\name{interacted_linear_estimators}
\alias{interacted_linear_estimators}
\title{Interacted linear regression models}
\usage{
interacted_linear_estimators(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  control.formula = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)
If siteID passed, it will weight the RA blocks within site and then average
these site estimates.

SEs come from the overall variance-covariance matrix.}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}
}
\value{
Dataframe of the different versions of this estimator (person and
  site weighted)
}
\description{
These linear models have block by treatment interaction terms.  The final ATE
estimates are then weighted average of the block (site) specific ATE
estimates.
}
