% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisiteresult.R
\name{is.multisiteresult}
\alias{is.multisiteresult}
\alias{multisiteresult}
\title{Is object a multisiteresult object?}
\usage{
is.multisiteresult(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a multisiteresult object?
}
\seealso{
Other multisiteresult: 
\code{\link{as.data.frame.multisiteresult}()},
\code{\link{print.multisiteresult}()}
}
\concept{multisiteresult}
