% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{is.qstatresult}
\alias{is.qstatresult}
\alias{qstatresult}
\title{Is object a qstatresult object?}
\usage{
is.qstatresult(x)
}
\arguments{
\item{x}{the object to check.}
}
\description{
Is object a qstatresult object?
}
\seealso{
Other qstatresult: 
\code{\link{as.data.frame.qstatresult}()},
\code{\link{print.qstatresult}()}
}
\concept{qstatresult}
