% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{linear_model_estimators}
\alias{linear_model_estimators}
\title{Estimate a series of linear models using different weighting schemes and
standard errors.}
\usage{
linear_model_estimators(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  block.stats = NULL,
  control.formula = NULL,
  weight.LM.method = "survey",
  weight.LM.scale.weights = TRUE
)
}
\value{
Data frame of the various results.
}
\description{
Estimate a series of linear models using different weighting schemes and
standard errors.
}
