% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_formula_utilities.R
\name{make.base.formula}
\alias{make.base.formula}
\title{Make a canonical fixed effect formula, possibly with control variables.}
\usage{
make.base.formula(Yobs = "Yobs", Z = "Z", control.formula = NULL, data = NULL)
}
\arguments{
\item{data}{}
}
\value{
Something like "Yobs ~ 1 + Z" or "Yobs ~ 1 + Z + X"
}
\description{
Make a canonical fixed effect formula, possibly with control variables.
}
