% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_formula_utilities.R
\name{make.canonical.data}
\alias{make.canonical.data}
\title{Check formula and control.formula for syntax}
\usage{
make.canonical.data(formula, control.formula = NULL, siteID = NULL, data)
}
\value{
Dataset with now-canonical variable names
}
\description{
Checks syntax and also presence of variables in the dataframe.
}
\details{
Then makes the outcome, treatment and block variables have canonical "Yobs", "Z", and "B" names.
}
