% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_formula_utilities.R
\name{make.FE.int.formula}
\alias{make.FE.int.formula}
\title{Make a canonical fixed effect formula, possibly with control variables.}
\usage{
make.FE.int.formula(Yobs = "Yobs", Z = "Z", B = "B",
  control.formula = NULL, data = NULL)
}
\arguments{
\item{data}{}
}
\value{
Something like "Yobs ~ 0 + Z * B - Z" or "Yobs ~ 0 + Z * B - Z + X"
}
\description{
Make a canonical fixed effect formula, possibly with control variables.
}
