% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_formula_utilities.R
\name{make_base_formula}
\alias{make_base_formula}
\title{Make a canonical fixed effect formula, possibly with control variables.}
\usage{
make_base_formula(Yobs = "Yobs", Z = "Z", control.formula = NULL, data = NULL)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{Name of treatment variable (assumed to exist in data)}

\item{control.formula}{What variables to control for, in the form of "~ X1 + X2".}

\item{data}{Dataframe holding all variables to be used in formula.}

\item{B}{Name of blocking variable (assumed to exist in data)}
}
\value{
Something like "Yobs ~ 1 + Z" or "Yobs ~ 1 + Z + X1 + X2"
}
\description{
Make a canonical fixed effect formula, possibly with control variables.
}
\keyword{internal}
