% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocks.R
\name{make_blocks}
\alias{make_blocks}
\title{Make blocks out of a continuous covariate by cutting it into pieces that are
ideally relatively homogenous.}
\usage{
make_blocks(X, method = c("small", "pair", "big", "none"), num.blocks)
}
\arguments{
\item{X}{covariate vector to block on}

\item{method}{How to block.}

\item{num.blocks}{If method is small, how many blocks to attempt to make}
}
\value{
Vector with one element per element of `X`
}
\description{
Given a dataset, create a bunch of blocks based on passed covariate and
return the factor of blocks
}
