% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{make_data}
\alias{make_data}
\alias{make_obs_data}
\title{Make simulated dataset from a list of block sizes}
\usage{
make_data(
  n_k,
  sigma_alpha = 1,
  sigma_tau = 0,
  tau = 5,
  sigma_0 = 1,
  sigma_1 = 1,
  corr = 0.5,
  exact = FALSE
)

make_obs_data(n_k = c(2, 3, 4, 8), p = 0.5, ...)
}
\arguments{
\item{n_k}{List of block sizes}

\item{sigma_alpha}{Standard deviation of the separation of the block mean Y0}

\item{sigma_tau}{Standard deviation of the separation of the block mean
treatment effects (Y1-Y0)}

\item{tau}{Cross site VARIANCE of site-level ATEs.}

\item{sigma_0}{Standard deviation of residual Y0 added to block means (can be
vector for individual variances per block)}

\item{sigma_1}{As `sigma_0` but for Y1s.}

\item{corr}{Correlation of Y0, Y1 within a block (can be vector of length K
for different blocks)}

\item{exact}{Passed to mvrnorm to control how block means are generated.}

\item{p}{Proportion of units treated (as close as possible given block
sizes).  This can be a vector with a probability for each block.}

\item{...}{Parameters to be passed to make_data()}
}
\value{
Dataframe with block indicators, Y0, and Y1.

Dataframe with original potential outcomes and observed outcome based
  on random assigment.
}
\description{
This method is the one that generates the simulation data used in Pashley &
Miratrix.

Generate data, form_blocks_from_continuous, and randomize within block and generate observed
potential outcomes
}
\details{
The block means are sampled from a multivariate normal distribution.  This
can be controlled so the variances are exact using the `exact` flag.
}
