% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plot.R
\name{make_diagnostic_plot}
\alias{make_diagnostic_plot}
\title{Make a diagnostic plot}
\usage{
make_diagnostic_plot(Y, Z, B, data = NULL)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors}
}
\description{
Plot block-specific variance estimates versus size of treatment group.
}
