% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{method_characteristics}
\alias{method_characteristics}
\title{Get list of methods in package along with characteristics of those methods}
\usage{
method_characteristics()
}
\value{
A tibble of characteristics.
}
\description{
Each method name comes with whether it targets average impact for individuals
or average impact of sites, and also whether it appears to be a finite-sample
method (assuming fixed sites, even if individuals are sampled within site) or
superpopulation method (assuming sites/blocks are themselves sampled).
}
