% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{paired_var}
\alias{paired_var}
\title{Matched pairs type variance function}
\usage{
paired_var(data_table, weighted = TRUE)
}
\arguments{
\item{data_table}{data frame containing info for set of blocks of same size}

\item{weighted}{indicates whether variance should be weighted by number of units}
}
\description{
Function that calculates matched pairs type variance.
}
