% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisiteresult.R
\name{print.multisiteresult}
\alias{print.multisiteresult}
\title{Pretty print result from Q-statistic call}
\usage{
\method{print}{multisiteresult}(x, ...)
}
\arguments{
\item{x}{A multisiteresult object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print result from Q-statistic call
}
\seealso{
Other multisiteresult: 
\code{\link{as.data.frame.multisiteresult}()},
\code{\link{is.multisiteresult}()}
}
\concept{multisiteresult}
