% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{print.qstatresult}
\alias{print.qstatresult}
\title{Pretty print result from Q-statistic call}
\usage{
\method{print}{qstatresult}(x, ...)
}
\arguments{
\item{x}{A qstatresult object.}

\item{...}{No extra options passed.}
}
\description{
Pretty print result from Q-statistic call
}
\seealso{
Other qstatresult: 
\code{\link{as.data.frame.qstatresult}()},
\code{\link{is.qstatresult}()}
}
\concept{qstatresult}
