% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{print.var_dat}
\alias{print.var_dat}
\title{Print method for fitdata output}
\usage{
\method{print}{var_dat}(x, ...)
}
\arguments{
\item{x}{output from fitdata (class=var_data)}

\item{...}{further arguments to match print}
}
\description{
Function that compares difference variance estimators for blocked designs.
}
