% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{rerandomize.data}
\alias{rerandomize.data}
\title{Rerandomize a given multisite simulated dataset and recalulate observed
outcomes}
\usage{
rerandomize.data(dat)
}
\arguments{
\item{dat}{Dataframe from, e.g., gen.dat()}
}
\value{
Same dataframe with treatment shuffled and Yobs recalculated.
}
\description{
Useful for simulations of finite sample inference where the dataset should be
held static and only randomization is necessary.
}
