% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{s.tc.func}
\alias{s.tc.func}
\title{Function that calculates variance of treatment effects.}
\usage{
s.tc.func(tau_vec)
}
\arguments{
\item{tau_vec}{vector of treatment effects}
}
\description{
Function that helps calculate bias by caculating the tue variance of treatment effects.
}
