% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_oracle.R
\name{s_tc_func}
\alias{s_tc_func}
\title{Calculates variance of treatment effects.}
\usage{
s_tc_func(tau_vec)
}
\arguments{
\item{tau_vec}{vector of treatment effects}
}
\description{
Function that helps calculate bias by caculating the true variance of treatment effects.
}
