% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{weighted.linear.estimators.naive}
\alias{weighted.linear.estimators.naive}
\title{Individual Weighted regression models using the lm and precision weighting}
\usage{
weighted.linear.estimators.naive(dat)
}
\description{
WARNING: This is the wrong kind of regression.  The weights are correct, but
the use is wrong.
}
