% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{weighted.linear.estimators}
\alias{weighted.linear.estimators}
\title{Survey-weighted adjusted linear regression}
\usage{
weighted.linear.estimators(Yobs, Z, B, data = NULL, siteID = NULL)
}
\arguments{
\item{data}{Dataframe to analyze, with Y, Z, and B columns.}
}
\value{
Dataframe of results for different estimators.
}
\description{
Use survey weight regression to reweight blocks to target unbiased ATE estimators.
}
