% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_linear_estimators.R
\name{weighted_linear_estimators}
\alias{weighted_linear_estimators}
\title{Survey-weighted adjusted linear regression}
\usage{
weighted_linear_estimators(
  formula,
  control.formula = NULL,
  siteID = NULL,
  data,
  scaled.weights = TRUE,
  weight.method = c("survey", "precision")
)
}
\arguments{
\item{formula}{TOADD}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}

\item{siteID}{Vector of site IDs if there are randomization blocks nested in}

\item{data}{Dataframe to analyze, with Y, Z, and B columns.}

\item{scaled.weights}{Logical TOADD}

\item{weight.method}{TOADD}
}
\value{
Dataframe of results for different estimators.
}
\description{
Use survey weight regression to reweight blocks to target unbiased ATE estimators.
}
