% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_support_functions.R
\name{within_blk_var}
\alias{within_blk_var}
\title{Variance function multiplied by n-1.}
\usage{
within_blk_var(Y)
}
\arguments{
\item{Y}{vector of potential outcomes}
}
\description{
Function that calculates scaled variance to help with bias calculation.
}
