% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIRC_method.R
\name{analysis.FIRC}
\alias{analysis.FIRC}
\title{Obtain p-value from testing for treatment heterogeniety with the FIRC model.}
\usage{
analysis.FIRC(data, REML = FALSE, anova = FALSE, pool = FALSE)
}
\description{
Utility function for simulation studies.  Simply calls estimate.ATE.FIRC and extracts p-value
}
