% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{analysis.idio}
\alias{analysis.idio}
\title{This file contains functions implementing the various methods for detecting
and estimating treatment effect variation across sites in a multi-site trial
using random-slope, random-intercept approaches.}
\usage{
analysis.idio(df)
}
\description{
We have:
  * Ideosyncratic test
  * Testing for systematic covariate (two versions, one with random coeff and one not)
  * Hybrid test doing a likelihood ratio of the hybrid model vs. no-variation model
}
\details{
Test for cross site variation using a liklihood ratio test vs. a model with no random slope (but random intercept).
This is the RIRC version of such a test.
}
