% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstatistic_method.R
\name{analysis.Qstatistic}
\alias{analysis.Qstatistic}
\title{Weiss et al. Q-statistic test
Dataframe has variables of
- B (site id)
- Yobs (outcome)
- Z (binary treatment 0/1)}
\usage{
analysis.Qstatistic(Yobs, Z, B, data = NULL)
}
\description{
Weiss et al. Q-statistic test
Dataframe has variables of
- B (site id)
- Yobs (outcome)
- Z (binary treatment 0/1)
}
