% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstatistic_method.R
\name{analysis.Qstatistic.extended}
\alias{analysis.Qstatistic.extended}
\alias{analysis.Qstatistic}
\title{Weiss et al. Q-statistic test
Dataframe has variables of
- sid (site id)
- Yobs (outcome)
- Z (binary treatment 0/1)}
\usage{
analysis.Qstatistic.extended(df)

analysis.Qstatistic(df)
}
\description{
Weiss et al. Q-statistic test
Dataframe has variables of
- sid (site id)
- Yobs (outcome)
- Z (binary treatment 0/1)

Utility wrapper function to just get the p-value without all the other stuff.
}
