% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_FIRC_cov.R
\name{analysis_FIRC_cov}
\alias{analysis_FIRC_cov}
\title{Covariate adjusted test for cross-site variation}
\usage{
analysis_FIRC_cov(
  Yobs,
  Z,
  B,
  X,
  siteID = NULL,
  data = NULL,
  REML = FALSE,
  anova = FALSE
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{X}{Site-level covariate ideally predictive of treatment variation.}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{Dataframe with defined Yobs, Z, and B variables.}

\item{REML}{TOADD}

\item{anova}{Use the anova() method to do the test for significance between the models. FALSE means do the modified chi-squared test.}
}
\description{
This method fits a FIRC model but also includes a site-level covariate, X,
potentially predictive of treatment impact.
}
\details{
This fits unpooled FIRC models.
}
