% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combination.R
\name{analysis_functions}
\alias{analysis_functions}
\alias{analysis_idio}
\alias{analysis_combination}
\alias{analysis_systematic}
\alias{analysis_systematic_RTx}
\title{Multilevel model based tests for cross site variation}
\usage{
analysis_idio(df)

analysis_combination(df)

analysis_systematic(df)

analysis_systematic_RTx(df)
}
\arguments{
\item{df}{Dataframe to analyze.  Columns of \code{Z}, \code{sid},
\code{Yobs}, and \code{W} assumed to exist, with those names.}
}
\value{
P-value from the test.
}
\description{
These minimal methods test for cross-site variation in
  a multi-site randomized trial using multilevel models. They
  assume a passed dataframe with specifically named columns.

  In particular, Yobs is the outcome column, Z is the treatment
  column, W is a site-level covariate column, and sid is the site
  ID column (categorical).
}
\section{Functions}{
\itemize{
\item \code{analysis_idio()}: The idiosyncratic version tests for cross site variation using a likelihood
ratio test vs. a model with no random slope (but random intercept). This is
the RIRC version of such a test.

\item \code{analysis_combination()}: This is a combination test: it uses a liklihood ratio test on a
model with both systematic and idiosyncratic variation (i.e., it
has an interaction of the covariate and outcome, and also
includes a random effect for treatment), comparing to a model
which has neither.

\item \code{analysis_systematic()}: Test for cross site variation by testing for a covariate
predictive of variation. This version does not allow for random
tx variation (random slope).

\item \code{analysis_systematic_RTx()}: Systematic test with the random idiosyncratic variation.  This
tests for systematic variation, ignoring any explicitly modeled
random (idiosyncratic) treatment variation.

}}
