% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combination.R, R/analysis_idio.R,
%   R/analysis_systematic.RTx.R
\name{analysis_combination}
\alias{analysis_combination}
\alias{analysis_idio}
\alias{analysis_systematic.RTx}
\title{Test for treatment variation across site.}
\usage{
analysis_combination(df)

analysis_idio(df)

analysis_systematic.RTx(df)
}
\arguments{
\item{df}{Dataframe to analyze.}
}
\description{
This is a combination test: it uses
a liklihood ratio test on model with both systematic and ideosyncratic
variation
}
