% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{analysis_Qstatistic}
\alias{analysis_Qstatistic}
\title{Calculate confidence interval for cross site variation using Q-statistic test inversion.}
\usage{
analysis_Qstatistic(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  alpha = 0.05,
  calc.CI = FALSE
)
}
\arguments{
\item{Yobs}{(outcome)}

\item{Z}{(binary treatment 0/1)}

\item{B}{(site id)}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified by siteID).}

\item{alpha}{The level of the test.  The CI will be a 1-2alpha confidence interval.}

\item{calc.CI}{Logical}
}
\description{
Code taken from prior work of Catherine Armstrong.
}
