% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_Qstatistic.R
\name{analysis_Qstatistic_stat}
\alias{analysis_Qstatistic_stat}
\alias{analysis_Qstatistic}
\title{Calculate confidence interval for cross site variation using Q-statistic test
inversion.}
\usage{
analysis_Qstatistic_stat(
  ATE_hat,
  SE_hat,
  alpha = 0.05,
  calc_CI = FALSE,
  verbose = FALSE,
  tolerance = 1e-05
)

analysis_Qstatistic(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  alpha = 0.05,
  calc_CI = FALSE
)
}
\arguments{
\item{ATE_hat}{List of estimated ATEs, one for each site.}

\item{SE_hat}{List of associated estimated SEs, assumed to be known.}

\item{alpha}{The level of the test.  The CI will be a 1-2alpha confidence
interval.}

\item{calc_CI}{Logical}

\item{verbose}{Print out a few diagnostics as the CI is being calculated, for
those curious.}

\item{tolerance}{For optimization and search for CI endpoints, how close to
optimal do we want?}

\item{Yobs}{(outcome)}

\item{Z}{(binary treatment 0/1)}

\item{B}{(site id)}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified
by siteID).}
}
\value{
List of several estimated quantities: cross-site variation, p-value
  for any variation, whether the p-value is less than passed alpha, the Q
  statistic itself, and the confidence interval range.
}
\description{
Two versions of this, one that takes raw data (in which case a linear model
is fit to it to get individual site impact estimates), and a second, '_stat',
that takes precalculated point estimates and standard errors.
}
\details{
The first simply calculates ATE_hat and SE_hat, and then calls the second.

Code based on prior work of Catherine Armstrong.  Augmented to do binary
search for endpoints for the test inversion procedure to generate confidence
intervals. Also calls optim() to find the point estimate by maximizing
p-value.
}
\examples{
analysis_Qstatistic_stat( rnorm(30, sd=0.1), rep(0.1,30), calc_CI = TRUE )
}
