% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstatistic_method.R
\name{analysis_Qstatistic_X}
\alias{analysis_Qstatistic_X}
\title{Weiss et al. Q-statistic test when there is an individual level covariate X}
\usage{
analysis_Qstatistic_X(data)
}
\description{
TO DO: Is the above description correct?
}
