% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blkvar.R
\docType{package}
\name{blkvar}
\alias{blkvar}
\title{\code{blkvar} package}
\description{
Estimating ATE and Cross-Site Variation for Multisite (block) RCTs
}
\details{
This code bundle is a package for estimating impacts and cross-site variation
in multi-site or blocked RCTs.

It is the compainion package for Pashley & Miratrix (2020) and Miratrix,
Weiss & Henderson (2020).

It also has a bunch of routines for running multisite simulation scenarios.
It is used in a variety of research projects connected with the Miratrix
CARES Lab (\url{https://cares.gse.harvard.edu/}) and other research that
Miratrix is part of.

The github has development versions, with more experimental code. For working
(cleaner) versions see CRAN. See
\href{https://github.com/lmiratrix/blkvar#readme}{GitHub} for current
development version.
}
\references{
Pashley & Miratrix (2020) Insights on Variance Estimation for
Blocked and Matched Pairs Designs on \url{https://arxiv.org/abs/1710.10342}

Miratrix, Weiss and Henderson (2020) "An Applied Researcher’s Guide to
Estimating Effects from Multisite Individually Randomized Trials: Estimands,
Estimators, and Estimates"
}
