% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dat_model.R
\name{block_distn}
\alias{block_distn}
\title{Generate site sizes for simulation of multisite data}
\usage{
block_distn(J, n.bar, size.ratio)
}
\arguments{
\item{J}{number of sites to generate.}

\item{n.bar}{Average site size.}

\item{size.ratio}{Index of}
}
\value{
Numeric vector of site sizes.
}
\description{
Generate site sizes for simulation of multisite data
}
\examples{
block_distn( 4, 10, 1 )
}
