% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{calc.summary.stats}
\alias{calc.summary.stats}
\title{Summarise data by block.}
\usage{
calc.summary.stats(Yobs, Z, B, data = NULL, siteID = NULL, add.neyman = FALSE)
}
\arguments{
\item{siteID}{If not null, name of siteID that has randomization blocks
nested inside.}

\item{dat}{Dataframe with defined Yobs, Z, and B variables.}
}
\value{
dataframe with summary statistics by block
}
\description{
Given dataframe or list of variables, return table of stats for each
randomization block.
}
