% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_stats.R
\name{calc_summary_stats_formula}
\alias{calc_summary_stats_formula}
\title{Summarise data by block.}
\usage{
calc_summary_stats_formula(
  formula,
  data = NULL,
  siteID = NULL,
  add.neyman = FALSE
)
}
\arguments{
\item{formula}{Formula of form Yobs ~ Z * B (ONLY)}

\item{data}{Dataframe with defined Yobs, Z, and B variables.}

\item{siteID}{If not null, name of siteID that has randomization blocks
nested inside.}

\item{add.neyman}{If TRUE, add block-specific SEs using Neyman formula of
$s^2_T/n_t + s^2_C/n_c$ as column in output.}
}
\value{
dataframe with summary statistics by block
}
\description{
Given dataframe or list of variables, return table of stats for each
randomization block.
}
\seealso{
Other calc_summary_stats: 
\code{\link{calc_summary_stats}()}
}
\concept{calc_summary_stats}
