% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_stats_formula.R
\name{calc_summary_stats_formula}
\alias{calc_summary_stats_formula}
\title{Summarise data by block.}
\usage{
calc_summary_stats_formula(
  formula,
  data = NULL,
  siteID = NULL,
  add.neyman = FALSE
)
}
\arguments{
\item{formula}{Formula of form Yobs ~ Z * B (ONLY)}

\item{data}{Dataframe with defined Yobs, Z, and B variables.}

\item{siteID}{if blocks B nested in sites, then pass the site indicator.}

\item{add.neyman}{TOADD}
}
\value{
dataframe with summary statistics by block
}
\description{
Given dataframe or list of variables, return table of stats for each
randomization block.
}
