% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary_stats_oracle.R
\name{calc_summary_stats_oracle}
\alias{calc_summary_stats_oracle}
\title{Summarise simulation data by block (where we know both Y0 and Y1)}
\usage{
calc_summary_stats_oracle(data, Y0 = "Y0", Y1 = "Y1", Z = "Z", B = "B")
}
\arguments{
\item{data}{Dataframe with defined Y0, Y1, and B variables.}

\item{Y0}{name of Y0 column}

\item{Y1}{name of Y1 column}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{B}{block ids}
}
\value{
dataframe with summary statistics by block
}
\description{
Summarise simulation data by block (where we know both Y0 and Y1)
}
