% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_methods}
\alias{compare_methods}
\title{Block variance method comparison function}
\usage{
compare_methods(Y, Z, B, data = NULL, include.block = TRUE,
  include.MLM = TRUE, include.RCTYes = TRUE, include.LM = TRUE,
  include.RCTYesBlended = FALSE)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors.  The columns
of the data matrix have to be in the order of Y, Z, B as column 1, 2, 3.}

\item{include.MLM}{Include MLM estimators}

\item{include.RCTYes}{Include RCTYes estimators}

\item{include.LM}{Include Linear Model-based estimators (including
Huber-White SEs, etc.)}

\item{include.RCTYesBlended}{Include RCTYes estimator applied to small block
and classic Neyman to large blocks.}
}
\description{
This function calculates the point estimates and SE estimates for a variety
of the blocked designs.
}
