% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_oracle.R
\name{compare_methods_oracle}
\alias{compare_methods_oracle}
\title{Function to compare estimators using the true values (i.e., whole science
table).  This is for simulaton studies where we know all the potential
outcomes.}
\usage{
compare_methods_oracle(Y, Z, B, p.mat, data = NULL)
}
\arguments{
\item{Y}{vector of all outcomes}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{B}{block ids}

\item{p.mat}{matrix with first column B,second column prop treated in
that block, p}

\item{data}{alternatively is matrix of Y,Z,B}
}
\description{
Function that returns some variance function estimates based on all potential
outcomes.
}
