% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_oracle.R
\name{compare_methods_oracle}
\alias{compare_methods_oracle}
\title{Compare estimators using the true values (i.e., full schedule of potential outcomes)}
\usage{
compare_methods_oracle(Y0, Y1, B, data = NULL, p_mat = NULL)
}
\arguments{
\item{Y0}{Vector of control potential outcomes  (or name of column in data
holding same).}

\item{Y1}{Vector of treatment potential outcomes  (or name of column in data
holding same).}

\item{B}{block ids  (or name of column in data holding same).}

\item{data}{alternatively is matrix of Y,Z,B}

\item{p_mat}{matrix with two columns, one row per block.  First column is
the categorical covariate denoting blocks, with same levels as in B, above.
Second column is proportion treated in that block.}
}
\description{
This method is for simulaton studies where we know all the potential
outcomes.
}
\details{
Function that returns some variance function estimates based on all potential
outcomes.
}
