% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods_variation.R
\name{compare_methods_variation}
\alias{compare_methods_variation}
\title{Compare different estimates of cross site variation.}
\usage{
compare_methods_variation(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  include_testing = TRUE,
  include_Q_estimate = TRUE,
  long_results = FALSE
)
}
\arguments{
\item{Yobs}{vector observed outcomes (or column name in data)}

\item{Z}{vector of assignment indicators (1==treated) (or column name in
data)}

\item{B}{vector of block ids (or column name in data)}

\item{siteID}{if blocks B nested in sites, then pass the site indicator.}

\item{data}{frame holding Y, Z, B and (possibly a column with name specified
by siteID).}

\item{include_testing}{Include tests for null of no cross site variation.}

\item{include_Q_estimate}{Include a estimate for tau_hat from Q method (involves calculating the full confidence interval so potentially time intensive).}

\item{long_results}{TRUE means each estimator gets a line in a data.frame.
FALSE gives all as columns in a 1-row dataframe.}
}
\description{
Given a dataframe, use the different methods to pull out point estimates for cross site variation and
(if desired) pvalues and return them all.
}
