% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisite_data_generators.R
\name{describe_data}
\alias{describe_data}
\title{Describe oracle dataset (for simulation studies)}
\usage{
describe_data(data, Y0 = "Y0", Y1 = "Y1", Z = "Z", sid = "sid")
}
\description{
Utility function to describe different characteristics of the distribution of
sites (given all the potential outcomes).
}
