% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe oracle dataset (for simulation studies)}
\usage{
describe_data(data, Y0 = "Y0", Y1 = "Y1", Z = "Z", sid = "sid")
}
\arguments{
\item{data}{Dataframe with defined Y0, Y1, and B variables.}

\item{Y0}{name of Y0 column}

\item{Y1}{name of Y1 column}

\item{Z}{vector that indicates if outcome is under treatment or control}

\item{sid}{block ids}
}
\description{
Utility function to describe different characteristics of the distribution of
sites (given all the potential outcomes).
}
