% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_method.R
\name{estimate.ATE.design.based.adjusted}
\alias{estimate.ATE.design.based.adjusted}
\title{Adjusted design based estimator for ATE}
\usage{
estimate.ATE.design.based.adjusted(
  formula,
  control.formula,
  data,
  siteID = NULL,
  method = c("finite", "superpop", "superpop.adj"),
  weight = c("individual", "site")
)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
Given two covariates, X1 and X2, calculate adjusted ATE estimates for the
multisite trial.  This uses the formula from the Schochet RCT Yes technical
document.
}
