% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{estimate.ATE.RIRC.pool}
\alias{estimate.ATE.RIRC.pool}
\title{Fit Random-intercept, random-slope model}
\usage{
estimate.ATE.RIRC.pool(Yobs, Z, sid, data = NULL, REML = FALSE,
  include.testing = TRUE)
}
\description{
This is a wrapper for a simpler lmer call.
}
