% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{estimate.ATE.RIRC}
\alias{estimate.ATE.RIRC}
\title{Fit Random-intercept, random-slope model}
\usage{
estimate.ATE.RIRC(Yobs, Z, sid, data = NULL, REML = FALSE)
}
\description{
This is a wrapper for a simpler lmer call.
}
