% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_method.R
\name{estimate.ATE.RIRC}
\alias{estimate.ATE.RIRC}
\title{Estimate the ATE using Random-Intercept, Random-Coefficient (RIRC) Models}
\usage{
estimate.ATE.RIRC(Yobs, Z, sid, data = NULL, REML = FALSE,
  include.testing = TRUE)
}
\description{
Estimate the ATE using Random-Intercept, Random-Coefficient (RIRC) Models
}
