% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_based_method.R
\name{estimate_ATE_design_based_adjusted}
\alias{estimate_ATE_design_based_adjusted}
\title{Adjusted design based estimator for ATE}
\usage{
estimate_ATE_design_based_adjusted(
  formula,
  control_formula,
  data,
  siteID = NULL,
  method = c("finite", "superpop", "superpop.adj"),
  weight = c("individual", "site", "tx")
)
}
\value{
Estimates of ATE along with SEs.
}
\description{
Given two covariates, X1 and X2, calculate adjusted ATE estimates for the
multisite trial.  This uses the formula from the Schochet RCT Yes technical
document.
}
