% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ATE_FIRC.R
\name{estimate_ATE_FIRC}
\alias{estimate_ATE_FIRC}
\title{Fit the FIRC model to estimate (1) ATE across sites and (2) cross site
treatment variation.}
\usage{
estimate_ATE_FIRC(
  Yobs,
  Z,
  B,
  siteID = NULL,
  control.formula = NULL,
  data = NULL,
  REML = FALSE,
  include.testing = TRUE,
  anova = FALSE,
  pool = FALSE
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{Name of the block indicator.}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}

\item{data}{Dataframe with all needed variables.}

\item{REML}{TOADD}

\item{include.testing}{Logical TOADD}

\item{anova}{Use the anova() method to do the test for significance between
the models.  FALSE means do the modified chi-squared test.}

\item{pool}{TRUE means tx and co have same reBual variance. FALSE gives seperate estimates for each (recommended, default).}
}
\description{
Acknowledgement: Unpooled version taken and adapted from Catherine's weiss.tau() method.
}
