% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIRC_method.R
\name{estimate_ATE_FIRC}
\alias{estimate_ATE_FIRC}
\title{Fit the FIRC model to estimate (1) ATE across sites and (2) cross site
treatment variation.}
\usage{
estimate_ATE_FIRC(
  Yobs,
  Z,
  B,
  siteID = NULL,
  control.formula = NULL,
  data = NULL,
  REML = FALSE,
  include.testing = TRUE,
  anova = FALSE,
  pool = FALSE
)
}
\arguments{
\item{B}{Name of the block indicator.}

\item{siteID}{Character name of the ID variable of site (blocks are conBered nested in site).  If omitted, then blocks are considered sites (the default).}

\item{anova}{Use the anova() method to do the test for significance between
the models.  FALSE means do the modified chi-squared test.}

\item{pool}{TRUE means tx and co have same reBual variance. FALSE gives seperate estimates for each (recommended, default).}
}
\description{
Acknowledgement: Unpooled version taken and adapted from Catherine's weiss.tau() method.
}
