% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ATE_RICC.R, R/estimate_ATE_RIRC.R
\name{estimate_ATE_RICC}
\alias{estimate_ATE_RICC}
\alias{estimate_ATE_RIRC}
\title{Estimate the ATE using Random-Intercept, Constant-Coefficient (RICC) Model.}
\usage{
estimate_ATE_RICC(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  control.formula = NULL
)

estimate_ATE_RIRC(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  include_testing = TRUE,
  pool = FALSE,
  control.formula = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{Name of the block indicator.}

\item{data}{Dataframe with all needed variables.}

\item{REML}{Logical, Restricted maximum likelihood or maximum likelihood estimation.}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}

\item{include_testing}{Logical Include likelihood ratio test for cross-site treatment variation.}

\item{pool}{TRUE means tx and co have same reBual variance. FALSE gives seperate estimates for each (recommended, default).}
}
\description{
This model has a single treatment coefficient, and a random intercept for the
site control average. So it is analogous to a fixed effect model, but with a
random effect.
}
\details{
There is no test for cross site variation for this method, since we assume none.
}
