% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIRC_method.R
\name{estimate_ATE_RIRC}
\alias{estimate_ATE_RIRC}
\alias{estimate_ATE_RICC}
\title{Estimate the ATE using Random-Intercept, Random-Coefficient (RIRC) Models}
\usage{
estimate_ATE_RIRC(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  include.testing = TRUE,
  pool = FALSE,
  control.formula = NULL
)

estimate_ATE_RICC(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  control.formula = NULL
)
}
\arguments{
\item{B}{Name of the block indicator.}

\item{pool}{TRUE means tx and co have same reBual variance. FALSE gives seperate estimates for each (recommended, default).}
}
\description{
This model has a single treatment coefficient, and a random intercept for the
site control average. So it is analogous to a fixed effect model, but with a
random effect.
}
\details{
There is no test for cross site variation for this method, since we assume none.
}
