% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ATE_RIRC_pool.R
\name{estimate_ATE_RIRC_pool}
\alias{estimate_ATE_RIRC_pool}
\title{Fit Random-intercept, random-slope model}
\usage{
estimate_ATE_RIRC_pool(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  include_testing = TRUE,
  keep_EB_estimates = TRUE,
  control_formula = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{Name of the block indicator.}

\item{data}{Dataframe with all needed variables.}

\item{REML}{Logical, Restricted maximum likelihood or maximum
likelihood estimation.  Default of TRUE if include_testing =
FALSE and FALSE otherwise. Note that FIRC can fail badly in
estimating cross-site variation when REML=FALSE.}

\item{include_testing}{Logical Include likelihood ratio test for
cross-site treatment variation.}

\item{keep_EB_estimates}{TRUE means returned object has EB
estimates.  FALSE means do not keep them.}

\item{control_formula}{The control_formula argument must be of the
form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}
}
\description{
This is a wrapper for a simpler lmer call.
}
