% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ATE_RIRC_pool.R
\name{estimate_ATE_RIRC_pool}
\alias{estimate_ATE_RIRC_pool}
\title{Fit Random-intercept, random-slope model}
\usage{
estimate_ATE_RIRC_pool(
  Yobs,
  Z,
  B,
  data = NULL,
  REML = FALSE,
  include_testing = TRUE,
  control.formula = NULL
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{Name of the block indicator.}

\item{data}{Dataframe with all needed variables.}

\item{REML}{Logical, Restricted maximum likelihood or maximum likelihood estimation.}

\item{include_testing}{Logical Include likelihood ratio test for cross-site treatment variation.}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}
}
\description{
This is a wrapper for a simpler lmer call.
}
