% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_var_est_func.R
\name{fitdata}
\alias{fitdata}
\title{Block variance estimation function.}
\usage{
fitdata(Y, Z, B, data = NULL, method = c("hybrid_m", "hybrid_p",
  "plug_in_big"), throw.warnings = TRUE)
}
\arguments{
\item{Y}{vector observed outcomes}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{data}{matrix of Y, Z, B, as alternative to using vectors}

\item{method}{is method of variance estimation, defauly "hybrid_m"}
}
\description{
This function takes observed data and returns a treatment effect estimate,
variance estimate and some summary info about the block structure.
}
