% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{linear.model.estimators}
\alias{linear.model.estimators}
\title{Estimate a series of linear models using different weighting schemes and
standard errors.}
\usage{
linear.model.estimators(Yobs, Z, B, siteID = NULL, data = NULL,
  block.stats = NULL, control.formula = NULL)
}
\value{
Data frame of the various results.
}
\description{
Estimate a series of linear models using different weighting schemes and
standard errors.
}
