% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_estimators.R
\name{linear_model_estimators}
\alias{linear_model_estimators}
\title{Estimate a series of linear models using different weighting schemes and
standard errors.}
\usage{
linear_model_estimators(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  block.stats = NULL,
  control.formula = NULL,
  weight.LM.method = "survey",
  weight.LM.scale.weights = TRUE
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{TOADD}

\item{block.stats}{TOADD}

\item{control.formula}{The control formula argument must be of the form ~ X1 + X2 + ... + XN. (nothing on left hand side of ~)}

\item{weight.LM.method}{TOADD}

\item{weight.LM.scale.weights}{TOADD}
}
\value{
Data frame of the various results.
}
\description{
Estimate a series of linear models using different weighting schemes and
standard errors.
}
