% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_estimators.R
\name{linear_model_estimators}
\alias{linear_model_estimators}
\title{Estimate a series of linear models using different weighting schemes and
standard errors.}
\usage{
linear_model_estimators(
  Yobs,
  Z,
  B,
  siteID = NULL,
  data = NULL,
  block.stats = NULL,
  control_formula = NULL,
  weight_LM_method = "survey",
  weight_LM_scale_weights = TRUE
)
}
\arguments{
\item{Yobs}{Name of outcome variable (assumed to exist in data)}

\item{Z}{vector of assignment indicators (1==treated)}

\item{B}{block ids}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{Dataframe of the data to analyse.}

\item{block.stats}{Table of precomputed block-level statistics (optional, for
speed concerns; this gets precomputed in compare_methods).}

\item{control_formula}{The control_formula argument must be of the form ~ X1
+ X2 + ... + XN. (nothing on left hand side of ~)}

\item{weight_LM_method}{Argument passed to weight.method of
weighted_linear_estimators}

\item{weight_LM_scale_weights}{Argument passed to sclae.weights of
weighted_linear_estimators}
}
\value{
Data frame of the various results.
}
\description{
Estimate a series of linear models using different weighting schemes and
standard errors.
}
