% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{plug_in_big}
\alias{plug_in_big}
\title{Plug in type variance estimator}
\usage{
plug_in_big(data.small, data.big)
}
\arguments{
\item{data.small}{data frame containg info for small blocks}

\item{data.big}{data frame containg info for small blocks}
}
\description{
Function to calculate variance estimate using average big block variances as plug-ins.
}
