% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R
\name{print.var_dat}
\alias{print.var_dat}
\title{Print method for block_estimator output}
\usage{
\method{print}{var_dat}(x, ...)
}
\arguments{
\item{x}{output from block_estimator (class=var_data)}

\item{...}{further arguments to match print}
}
\value{
None. Side effect of printing to console.
}
\description{
Print out results from block_estimator() in a nice way.
}
