% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multilevel_data.R
\name{rerandomize_data}
\alias{rerandomize_data}
\title{Rerandomize a given multisite simulated dataset and recalulate observed
outcomes}
\usage{
rerandomize_data(dat)
}
\arguments{
\item{dat}{Dataframe from, e.g., generate_multilevel_data() that has two columns, 'sid', and 'Z'.}
}
\value{
Same dataframe with treatment shuffled and Yobs recalculated.
}
\description{
Permute the treatment vector Z within site ID.  Useful for simulations of
finite sample inference where the dataset should be held static and only
randomization is necessary.
}
