% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_experiment_method.R,
%   R/linear_model_method.R
\name{scat}
\alias{scat}
\title{Utility to help printing out nicely formatted stuff.}
\usage{
scat(str, ...)

scat(str, ...)
}
\description{
Utility to help printing out nicely formatted stuff.

Utility to help printing out nicely formatted stuff.
}
