% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{weighted_linear_estimators}
\alias{weighted_linear_estimators}
\title{Survey-weighted adjusted linear regression}
\usage{
weighted_linear_estimators(
  formula,
  control_formula = NULL,
  siteID = NULL,
  data,
  scaled.weights = TRUE,
  weight.method = c("survey", "precision")
)
}
\arguments{
\item{formula}{Specification of outcome, treatment, and block ID variables as a formula of form "Yobs ~ Z*B" (order of variables is important).}

\item{control_formula}{The control_formula argument must be of the form ~ X1
+ X2 + ... + XN. (nothing on left hand side of ~)}

\item{siteID}{If not null, name of siteID that has randomization blocks}

\item{data}{Dataframe of the data to analyse.}

\item{scaled.weights}{Logical Scale the weights by overall proportions of treated and control.}

\item{weight.method}{Use survey package (svgglm) or classic OLS (lm) for model fitting.}
}
\value{
Dataframe of results for different estimators.
}
\description{
Use survey weight regression to reweight blocks to target unbiased ATE estimators.
}
\seealso{
Other linear model estimators: 
\code{\link{fixed_effect_estimators}()},
\code{\link{interacted_linear_estimators}()},
\code{\link{linear_model_estimators}()}
}
\concept{linear model estimators}
