% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model_method.R
\name{weighted_linear_estimators}
\alias{weighted_linear_estimators}
\title{Survey-weighted adjusted linear regression}
\usage{
weighted_linear_estimators(
  formula,
  control.formula = NULL,
  siteID = NULL,
  data,
  scaled.weights = TRUE,
  weight.method = c("survey", "precision")
)
}
\arguments{
\item{data}{Dataframe to analyze, with Y, Z, and B columns.}
}
\value{
Dataframe of results for different estimators.
}
\description{
Use survey weight regression to reweight blocks to target unbiased ATE estimators.
}
